Hooked up the menu and cleaned up the code.
DeckOfCards- I deleted the unnecessary local variables.

GuessTheCardGame- I added a class which is going to be a game that makes the user guess if the next card is lower or higher. I was able to get the method Run() up and going which handles the menu and calling of the different user method choices.