Improved the method Play().
I changed the user unput choices to y or n. And now the user only has to guess if the next card will be higher or not.

I consolidated the ifs statements in Play().

Gave the methods in charge or verififying the validity of the user's input more distinctive names.

Created the local variable score.

Also corrected a few formatting issues.