Created a List of key values field which will be used to store the score board of the game.
Added an if statement in the constructor to check if a file for the score board exists, if not, one will be created and random scores and names will be added to it.

Created an method in charge of updating the score board with any new high scores.

Reorginized the code.